requirements.txt
requests
packaging
pycryptodome
python-dateutil
pprint
M2Crypto
